backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
emergentintegrations
transformers>=4.36.0
torch>=2.1.0
tweepy>=4.14.0
facebook-scraper>=0.2.59
instaloader>=4.10.3
imagehash>=4.3.1
Pillow>=10.0.0
textblob>=0.17.1
scikit-learn>=1.3.0
schedule>=1.2.0
asyncio-mqtt>=0.16.1